refactor(gateway): remove dead sessions.observer.ask rpc
