+++ apps/android/app/src/main/java/ai/openclaw/app/gateway/GatewayProtocol.kt
@@ -485,7 +485,6 @@ enum class GatewayMethod(
   BoardPromptAuthorize("board.prompt.authorize"),
   BoardDataRead("board.data.read"),
   BoardAction("board.action"),
-  SessionsObserverAsk("sessions.observer.ask"),
   SessionsObserverVisibility("sessions.observer.visibility"),
   SessionVisibilitySet("session.visibility.set"),
   SessionMembersList("session.members.list"),
