+++ apps/shared/OpenClawKit/Sources/OpenClawProtocol/GatewayModels.swift
@@ -5246,42 +5246,6 @@ public struct SessionsCompanionStateResult: Codable, Sendable {
     }
 }
 
-public struct SessionsObserverAskParams: Codable, Sendable {
-    public let sessionkey: String
-    public let question: String
-
-    public init(
-        sessionkey: String,
-        question: String)
-    {
-        self.sessionkey = sessionkey
-        self.question = question
-    }
-
-    private enum CodingKeys: String, CodingKey {
-        case sessionkey = "sessionKey"
-        case question
-    }
-}
-
-public struct SessionsObserverAskResult: Codable, Sendable {
-    public let answer: String
-    public let digestrevision: Int?
-
-    public init(
-        answer: String,
-        digestrevision: Int? = nil)
-    {
-        self.answer = answer
-        self.digestrevision = digestrevision
-    }
-
-    private enum CodingKeys: String, CodingKey {
-        case answer
-        case digestrevision = "digestRevision"
-    }
-}
-
 public struct SessionsObserverVisibilityParams: Codable, Sendable {
     public let visible: Bool
