+++ packages/gateway-protocol/src/gateway-error-details.ts
@@ -24,8 +24,6 @@ export const GatewayErrorDetailCodes = {
   MISSING_SCOPE: "MISSING_SCOPE",
   MCP_APP_VIEW_EXPIRED: "MCP_APP_VIEW_EXPIRED",
   SESSION_COMPANION_BUSY: "SESSION_COMPANION_BUSY",
-  SESSION_OBSERVER_BUSY: "SESSION_OBSERVER_BUSY",
-  SESSION_OBSERVER_UNAVAILABLE: "SESSION_OBSERVER_UNAVAILABLE",
   UNKNOWN_AGENT_ID: "UNKNOWN_AGENT_ID",
   WIZARD_NOT_FOUND: "WIZARD_NOT_FOUND",
 } as const;
