+++ packages/gateway-protocol/src/schema/protocol-schemas.ts
@@ -562,8 +562,6 @@ import {
   SessionsCompanionResetResultSchema,
   SessionsCompanionStateParamsSchema,
   SessionsCompanionStateResultSchema,
-  SessionsObserverAskParamsSchema,
-  SessionsObserverAskResultSchema,
   SessionsObserverVisibilityParamsSchema,
   SessionsObserverVisibilityResultSchema,
   SessionWorktreeInfoSchema,
@@ -869,8 +867,6 @@ export const ProtocolSchemas = {
   SessionsCompanionResetResult: SessionsCompanionResetResultSchema,
   SessionsCompanionStateParams: SessionsCompanionStateParamsSchema,
   SessionsCompanionStateResult: SessionsCompanionStateResultSchema,
-  SessionsObserverAskParams: SessionsObserverAskParamsSchema,
-  SessionsObserverAskResult: SessionsObserverAskResultSchema,
   SessionsObserverVisibilityParams: SessionsObserverVisibilityParamsSchema,
   SessionsObserverVisibilityResult: SessionsObserverVisibilityResultSchema,
   SessionVisibility: SessionVisibilitySchema,
