+++ src/gateway/methods/core-descriptors.since.test.ts
@@ -61,7 +61,6 @@ const CURRENT_TRAIN_METHODS = [
   "gateway.suspend.resume",
   "ui.command",
   "device.pair.rename",
-  "sessions.observer.ask",
   "sessions.observer.visibility",
   "sessions.companion.ask",
   "sessions.companion.state",
